pyproject.toml
[project]
name = "twr"
version = "0.1.0"
description = "Time-Weighted Return calculation system with PostgreSQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["psycopg2-binary>=2.9.9", "jinja2>=3.1.0"]

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = [
    "I",      # isort (import sorting)
    "F",      # pyflakes (includes unused imports)
    "ANN001", # missing-type-function-argument
    "ANN201", # missing-return-type-undocumented-public-function
    "ANN202", # missing-return-type-private-function
]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning:testcontainers.*:"]
pythonpath = ["src"]

[dependency-groups]
dev = [
  "plotly>=6.5.0",
  "pytest>=7.4.0",
  "testcontainers[postgres]",
  "watchdog>=6.0.0",
  "faker>=24.0.0",
  "streamlit>=1.52.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "rich>=14.2.0",
  "pudb>=2025.1.5",
  "ty>=0.0.17",
]
